Adds default targeting accessor and extension method WithTargeting variation for ASP.NET
