--- examples/FeatureFlagDemo/ClaimTypes.cs
@@ -1,10 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-//
-namespace FeatureFlagDemo
-{
-    static class ClaimTypes
-    {
-        public static string GroupName = "http://schemas.featureflagdemo.featuremanagement.microsoft.com/claims/groupname";
-    }
-}
